feat: force bundle LiveReload to prevent process.env variables not being replaced
