+++ packages/remix-react/index.tsx
@@ -36,7 +36,6 @@ export {
   NavLink,
   Form,
   PrefetchPageLinks,
-  LiveReload,
   useFormAction,
   useSubmit,
   useTransition,
